# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ParkIt API",
        "description": "API for managing parkings and authentication in the ParkIt application",
        "contact": {
            "name": "API Support",
            "email": "support@parkit.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com"
        }
    ],
    "paths": {
...
        "/api/places/{place}/reservations": {
            "post": {
                "tags": [
                    "Reservations"
                ],
                "summary": "Create a reservation",
                "operationId": "d142a0f8a2111867b7c4c24df3b04e44",
                "parameters": [
                    {
                        "name": "place",
                        "in": "path",
                        "description": "Place ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "start": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2023-10-01T10:00:00Z"
                                    },
                                    "end": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2023-10-01T12:00:00Z"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Reservation created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "user_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "place_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "start": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2023-10-01T10:00:00Z"
                                        },
                                        "end": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2023-10-01T12:00:00Z"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Place is already reserved"
                    }
                }
            }
        },
        "/api/reservations/{reservation}": {
            "get": {
                "tags": [
                    "Reservations"
                ],
                "summary": "Get a reservation by ID",
                "operationId": "2aefc564f94e483fb93920c2935eeafe",
                "parameters": [
                    {
                        "name": "reservation",
                        "in": "path",
                        "description": "Reservation ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Reservation details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "user_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "place_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "start": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2023-10-01T10:00:00Z"
                                        },
                                        "end": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2023-10-01T12:00:00Z"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Reservations"
                ],
                "summary": "Update a reservation",
                "operationId": "459cb3958996a0266b294be06a643db0",
                "parameters": [
                    {
                        "name": "reservation",
                        "in": "path",
                        "description": "Reservation ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "start": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2023-10-01T10:00:00Z"
                                    },
                                    "end": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2023-10-01T12:00:00Z"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Reservation updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "user_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "place_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "start": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2023-10-01T10:00:00Z"
                                        },
                                        "end": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2023-10-01T12:00:00Z"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "400": {
                        "description": "Place is already reserved"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Reservations"
                ],
                "summary": "Delete a reservation",
                "operationId": "9cb75e88bfb3a102881ea262dde33019",
                "parameters": [
                    {
                        "name": "reservation",
                        "in": "path",
                        "description": "Reservation ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Reservation deleted successfully"
                    },
                    "403": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
...
    },
...
}